feat(printf): add support for long format specifiers

+++ kernel/src/lib/printf.hpp
@@ -11,8 +11,8 @@ int printf(const char* format, ...);
 int vprintf(const char* format, va_list args);
 
 void print_number(int num, int base = 10, int width = 0, bool pad_zero = false);
-void print_unsigned(unsigned long num, int base = 10, int width = 0, bool pad_zero = false);
-void print_hex(unsigned long num, int width = 0, bool pad_zero = false);
+void print_unsigned(unsigned long long num, int base = 10, int width = 0, bool pad_zero = false);
+void print_hex(unsigned long long num, int width = 0, bool pad_zero = false);
 void print_pointer(const void* ptr);
 
 #ifdef __cplusplus
